requirements.txt
pandas>=2.0.0
numpy>=1.24.0
unidecode>=1.3.0
rapidfuzz>=3.0.0
openpyxl>=3.1.0
pyyaml>=6.0
streamlit>=1.30.0
